+++ packages/plugin-ext/src/plugin/languages/diagnostics.ts
@@ -249,7 +249,7 @@ export class Diagnostics {
     private static GENERATED_DIAGNOSTIC_COLLECTION_NAME_PREFIX = '_generated_diagnostic_collection_name_#';
 
     private proxy: LanguagesMain;
-    private diagnosticCollections: Map<string, DiagnosticCollection>; // id -> diagnostic colection
+    private diagnosticCollections: Map<string, DiagnosticCollection>; // id -> diagnostic collection
 
     private diagnosticsChangedEmitter = new Emitter<theia.DiagnosticChangeEvent>();
     public readonly onDidChangeDiagnostics: Event<theia.DiagnosticChangeEvent> = this.diagnosticsChangedEmitter.event;
@@ -264,7 +264,7 @@ export class Diagnostics {
     getDiagnostics(): [theia.Uri, theia.Diagnostic[]][];
     getDiagnostics(resource?: URI): theia.Diagnostic[] | [URI, theia.Diagnostic[]][] {
         if (resource) {
-            return this.getAllDiagnisticsForResource(resource);
+            return this.getAllDiagnosticsForResource(resource);
         } else {
             return this.getAllDiagnostics();
         }
@@ -291,7 +291,7 @@ export class Diagnostics {
         return v4();
     }
 
-    private getAllDiagnisticsForResource(uri: URI): theia.Diagnostic[] {
+    private getAllDiagnosticsForResource(uri: URI): theia.Diagnostic[] {
         let result: theia.Diagnostic[] = [];
         this.diagnosticCollections.forEach(diagnosticCollection => {
             const diagnostics = diagnosticCollection.get(uri);